GeoEMTIPDemonstration

Documentation for GeoEMTIPDemonstration.

GeoEMTIPDemonstration.DateInterval

GeoEMTIPDemonstration.InformationForFigure

GeoEMTIPDemonstration.prep202304!

GeoEMTIPDemonstration.set_aog_pallete!

GeoEMTIPDemonstration.textoffset

GeoEMTIPDemonstration.to_days

GeoEMTIPDemonstration.to_days

GeoEMTIPDemonstration.toindex

GeoEMTIPDemonstration.uniqsomething!

GeoEMTIPDemonstration.DateInterval — Method

Iterable DateInterval(edges); returns (t0, t1) from edges where (t0, t1) = edges[i], edges[i + 1] for i in 1:length(edges)-1.

Example

julia> using Dates

julia> edges = [Date(0000, 1, 1), Date(2017, 10, 1), Date(2020, 10, 1), Date(9999,12,31)];

julia> [(t0, t1) for (t0, t1) in DateInterval(edges)]
3-element Vector{Tuple{Date, Date}}:
 (Date("0000-01-01"), Date("2017-10-01"))
 (Date("2017-10-01"), Date("2020-10-01"))
 (Date("2020-10-01"), Date("9999-12-31"))

GeoEMTIPDemonstration.InformationForFigure — Type

GeoEMTIPDemonstration.prep202304! — Method

Given df = CSV.read("summary_test.csv", DataFrame), prep202304!(df) calculates the following new columns:

:FittingDegree

:prp_ind (for Makie plot use)

:frc_ind (for Makie plot use)

GeoEMTIPDemonstration.set_aog_pallete! — Method

You may alternatively do:

colors = ["#FC7808", "#8C00EC", "#107A78"]
style = (color= colors, )
df = (x=rand(["a", "b", "c"], 100), y=rand(100))
plt = data(df) * mapping(:x, :y, color = :x) * visual(BoxPlot)
draw(plt, palettes = style)

See the thread Styling AlgebraOfGraphics boxplots.

GeoEMTIPDemonstration.textoffset — Method

Example: Text offset according to align

station_location = DataFrame(
    Lon = [112.1, 112.3, 112.5],
    Lat = [21.1, 23.3, 25.5],
    code = ["YL", "AA", "KUOL"],
    TextAlign = [
        (:center, :top),
        (:right, :bottom),
        (:left, :bottom),
    ]
)
text(station_location.Lon, station_location.Lat;
    text=station_location.code,
    align=station_location.TextAlign,
    offset = textoffset.(station_location.TextAlign, 4),
    fontsize=15
    )

GeoEMTIPDemonstration.to_days — Method

to_days(p::Dates.CompoundPeriod) returns the summation of CompoundPeriod.

GeoEMTIPDemonstration.to_days — Method

Given p::Period, to_days(p) returns Real number of the value in the unit of day.

GeoEMTIPDemonstration.toindex — Method

toindex(str::AbstractString, uniqlabels::Vector{<:AbstractString}) returns the only index of uniqlabels occurring in str.

GeoEMTIPDemonstration.uniqsomething! — Method

uniqcol = uniqsomething!(df, col) add new column "col_ind" as the integer indices for a vector of string uniqcol.