GeoEMTIPDemonstration

Documentation for GeoEMTIPDemonstration.

GeoEMTIPDemonstration.binomial_cache

GeoEMTIPDemonstration.DateInterval

GeoEMTIPDemonstration.InformationForFigure

GeoEMTIPDemonstration.TrialTable

GeoEMTIPDemonstration.dtstr2nday

GeoEMTIPDemonstration.getdcb

GeoEMTIPDemonstration.prep202304!

GeoEMTIPDemonstration.to_days

GeoEMTIPDemonstration.to_days

GeoEMTIPDemonstration.toindex

GeoEMTIPDemonstration.uniqsomething!

GeoEMTIPDemonstration.binomial_cache — Constant

A dictionary to cache computed binomial coefficients.

GeoEMTIPDemonstration.DateInterval — Method

Iterable DateInterval(edges); returns (t0, t1) from edges where (t0, t1) = edges[i], edges[i + 1] for i in 1:length(edges)-1.

Example

julia> using Dates

julia> edges = [Date(0000, 1, 1), Date(2017, 10, 1), Date(2020, 10, 1), Date(9999,12,31)];

julia> [(t0, t1) for (t0, t1) in DateInterval(edges)]
3-element Vector{Tuple{Date, Date}}:
 (Date("0000-01-01"), Date("2017-10-01"))
 (Date("2017-10-01"), Date("2020-10-01"))
 (Date("2020-10-01"), Date("9999-12-31"))

GeoEMTIPDemonstration.InformationForFigure — Type

GeoEMTIPDemonstration.TrialTable — Type

The children of TrialTable are used to identify the matching between dataset and trial name. The matching methods are not general, and should be defined for specific projects; for example, see load_data.jl in the Project2024.

GeoEMTIPDemonstration.dtstr2nday — Method

Given dtstr in format "yyyymmdd-yyyymmdd", dtstr2nday(dtstr) returns the number of days covering this range. Saying the two dates are DateTime, dt0 and dt1, it should be noted that dtstr2nday returns (dt1 - dt0).value + 1 rather than (dt1 - dt0).value.

dtstr = "20150402-20150928"
dtstr2nday(dtstr)

GeoEMTIPDemonstration.getdcb — Method

getdcb(α, neq) calculate alarmed rate and hit rate for the given complementary confidence level α and number of earthquakes neq.

getdcb applies big if neq is larger than keyword argument applybig (40 by default).

Since julia is slow in handling BigInt, getdcb by default looks into dictionary binomial_cache first if there are cached results already; this behavior is controlled by cache=true keyword argument.

GeoEMTIPDemonstration.prep202304! — Method

Given df = CSV.read("summary_test.csv", DataFrame), prep202304!(df) calculates the following new columns:

:FittingDegree

:prp_ind (for Makie plot use)

:frc_ind (for Makie plot use)

GeoEMTIPDemonstration.to_days — Method

to_days(p::Dates.CompoundPeriod) returns the summation of CompoundPeriod.

GeoEMTIPDemonstration.to_days — Method

Given p::Period, to_days(p) returns Real number of the value in the unit of day.

GeoEMTIPDemonstration.toindex — Method

toindex(str::AbstractString, uniqlabels::Vector{<:AbstractString}) returns the only index of uniqlabels exactly matched by str.

GeoEMTIPDemonstration.uniqsomething! — Method

uniqcol = uniqsomething!(df, col) add new column "col_ind" as the integer indices for a vector of string uniqcol.